+++ src/calc.cpp
@@ -15,7 +15,7 @@ T Mult(T a, T b) {
     return (a * b);
 }
 
-int Div(float a, float b) {
+double Div(double a, double b) {
     if (b == 0.0) {
         throw 0;
     }
